# storage/engine: bring pebbleMVCCScanner closer to the C++ version