Miscellaneous changes to bring `pebbleMVCCScanner` in line with the C++
version.

```
name                                                    old time/op    new time/op     delta
MVCCScan_Pebble/rows=1/versions=1/valueSize=8-16          4.00µs ± 6%     2.58µs ± 6%  -35.51%  (p=0.000 n=10+10)
MVCCScan_Pebble/rows=1/versions=1/valueSize=64-16         6.16µs ± 2%     3.78µs ± 5%  -38.65%  (p=0.000 n=8+10)
MVCCScan_Pebble/rows=1/versions=1/valueSize=512-16        4.81µs ± 4%     3.15µs ± 8%  -34.57%  (p=0.000 n=10+10)
MVCCScan_Pebble/rows=10/versions=1/valueSize=8-16         5.96µs ± 9%     4.15µs ± 2%  -30.26%  (p=0.000 n=10+9)
MVCCScan_Pebble/rows=10/versions=1/valueSize=64-16        8.50µs ± 3%     6.02µs ± 3%  -29.22%  (p=0.000 n=10+10)
MVCCScan_Pebble/rows=10/versions=1/valueSize=512-16       8.46µs ± 6%     6.67µs ± 2%  -21.17%  (p=0.000 n=10+10)
MVCCScan_Pebble/rows=100/versions=1/valueSize=8-16        21.6µs ± 2%     16.8µs ± 2%  -22.50%  (p=0.000 n=10+10)
MVCCScan_Pebble/rows=100/versions=1/valueSize=64-16       29.5µs ± 3%     22.8µs ± 3%  -22.72%  (p=0.000 n=9+10)
MVCCScan_Pebble/rows=100/versions=1/valueSize=512-16      33.9µs ± 3%     26.6µs ± 2%  -21.46%  (p=0.000 n=10+10)
MVCCScan_Pebble/rows=1000/versions=1/valueSize=8-16        181µs ± 9%      136µs ± 1%  -25.20%  (p=0.000 n=10+9)
MVCCScan_Pebble/rows=1000/versions=1/valueSize=64-16       215µs ± 5%      165µs ± 2%  -23.12%  (p=0.000 n=10+10)
MVCCScan_Pebble/rows=1000/versions=1/valueSize=512-16      255µs ± 4%      216µs ± 1%  -15.13%  (p=0.000 n=10+10)
MVCCScan_Pebble/rows=10000/versions=1/valueSize=8-16      1.55ms ± 2%     1.22ms ± 2%  -21.50%  (p=0.000 n=9+10)
MVCCScan_Pebble/rows=10000/versions=1/valueSize=64-16     1.92ms ± 2%     1.53ms ± 5%  -20.39%  (p=0.000 n=10+10)
MVCCScan_Pebble/rows=10000/versions=1/valueSize=512-16    2.69ms ± 2%     2.19ms ± 1%  -18.33%  (p=0.000 n=10+8)
```

```
name                                        old time/op    new time/op    delta
MVCCGet_Pebble/versions=1/valueSize=8-16      3.29µs ± 3%    2.18µs ± 3%  -33.66%  (p=0.000 n=10+9)
MVCCGet_Pebble/versions=10/valueSize=8-16     4.05µs ± 2%    2.73µs ± 3%  -32.67%  (p=0.000 n=9+9)
MVCCGet_Pebble/versions=100/valueSize=8-16    15.2µs ± 7%     8.9µs ± 3%  -41.23%  (p=0.000 n=10+9)
```

The performance improvement is primarily due to avoiding an extra copy
of the key and value in `updateCurrent()`.

Release note: None